fix: Swagger Tag 수정
